+++ internal/repos/postgresql/work_item_project_demo_two.go
@@ -22,7 +22,7 @@ func NewDemoTwoWorkItem() *DemoTwoWorkItem {
 
 var _ repos.DemoTwoWorkItem = (*DemoTwoWorkItem)(nil)
 
-func (u *DemoTwoWorkItem) ByID(ctx context.Context, d db.DBTX, id int64, opts ...db.WorkItemSelectConfigOption) (*db.WorkItem, error) {
+func (u *DemoTwoWorkItem) ByID(ctx context.Context, d db.DBTX, id int, opts ...db.WorkItemSelectConfigOption) (*db.WorkItem, error) {
 	extraOpts := []db.WorkItemSelectConfigOption{db.WithWorkItemJoin(db.WorkItemJoins{DemoTwoWorkItem: true})}
 	return db.WorkItemByWorkItemID(ctx, d, id, (append(extraOpts, opts...))...)
 }
@@ -44,7 +44,7 @@ func (u *DemoTwoWorkItem) Create(ctx context.Context, d db.DBTX, params repos.De
 	return workItem, nil
 }
 
-func (u *DemoTwoWorkItem) Update(ctx context.Context, d db.DBTX, id int64, params repos.DemoTwoWorkItemUpdateParams) (*db.WorkItem, error) {
+func (u *DemoTwoWorkItem) Update(ctx context.Context, d db.DBTX, id int, params repos.DemoTwoWorkItemUpdateParams) (*db.WorkItem, error) {
 	workItem, err := u.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItem by id: %w", parseErrorDetail(err))
@@ -73,7 +73,7 @@ func (u *DemoTwoWorkItem) Update(ctx context.Context, d db.DBTX, id int64, param
 	return workItem, err
 }
 
-func (u *DemoTwoWorkItem) Delete(ctx context.Context, d db.DBTX, id int64) (*db.WorkItem, error) {
+func (u *DemoTwoWorkItem) Delete(ctx context.Context, d db.DBTX, id int) (*db.WorkItem, error) {
 	workItem, err := u.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItem: %w", parseErrorDetail(err))
@@ -87,7 +87,7 @@ func (u *DemoTwoWorkItem) Delete(ctx context.Context, d db.DBTX, id int64) (*db.
 	return workItem, err
 }
 
-func (u *DemoTwoWorkItem) Restore(ctx context.Context, d db.DBTX, id int64) (*db.WorkItem, error) {
+func (u *DemoTwoWorkItem) Restore(ctx context.Context, d db.DBTX, id int) (*db.WorkItem, error) {
 	var err error
 	workItem := &db.WorkItem{
 		WorkItemID: id,
