+++ internal/repos/reposwrappers/demotwoworkitem_with_prometheus.gen.go
@@ -39,7 +39,7 @@ func NewDemoTwoWorkItemWithPrometheus(base repos.DemoTwoWorkItem, instanceName s
 }
 
 // ByID implements repos.DemoTwoWorkItem
-func (_d DemoTwoWorkItemWithPrometheus) ByID(ctx context.Context, d db.DBTX, id int64, opts ...db.WorkItemSelectConfigOption) (wp1 *db.WorkItem, err error) {
+func (_d DemoTwoWorkItemWithPrometheus) ByID(ctx context.Context, d db.DBTX, id int, opts ...db.WorkItemSelectConfigOption) (wp1 *db.WorkItem, err error) {
 	_since := time.Now()
 	defer func() {
 		result := "ok"
@@ -67,7 +67,7 @@ func (_d DemoTwoWorkItemWithPrometheus) Create(ctx context.Context, d db.DBTX, p
 }
 
 // Delete implements repos.DemoTwoWorkItem
-func (_d DemoTwoWorkItemWithPrometheus) Delete(ctx context.Context, d db.DBTX, id int64) (wp1 *db.WorkItem, err error) {
+func (_d DemoTwoWorkItemWithPrometheus) Delete(ctx context.Context, d db.DBTX, id int) (wp1 *db.WorkItem, err error) {
 	_since := time.Now()
 	defer func() {
 		result := "ok"
@@ -81,7 +81,7 @@ func (_d DemoTwoWorkItemWithPrometheus) Delete(ctx context.Context, d db.DBTX, i
 }
 
 // Restore implements repos.DemoTwoWorkItem
-func (_d DemoTwoWorkItemWithPrometheus) Restore(ctx context.Context, d db.DBTX, id int64) (wp1 *db.WorkItem, err error) {
+func (_d DemoTwoWorkItemWithPrometheus) Restore(ctx context.Context, d db.DBTX, id int) (wp1 *db.WorkItem, err error) {
 	_since := time.Now()
 	defer func() {
 		result := "ok"
@@ -95,7 +95,7 @@ func (_d DemoTwoWorkItemWithPrometheus) Restore(ctx context.Context, d db.DBTX, 
 }
 
 // Update implements repos.DemoTwoWorkItem
-func (_d DemoTwoWorkItemWithPrometheus) Update(ctx context.Context, d db.DBTX, id int64, params repos.DemoTwoWorkItemUpdateParams) (wp1 *db.WorkItem, err error) {
+func (_d DemoTwoWorkItemWithPrometheus) Update(ctx context.Context, d db.DBTX, id int, params repos.DemoTwoWorkItemUpdateParams) (wp1 *db.WorkItem, err error) {
 	_since := time.Now()
 	defer func() {
 		result := "ok"
